Wireless LED RGB Light-IoT Based
In this mini-project, an RGB light-emitting diode ( lamp was designed for a smart
home system. The ultimate aim of this project is to produce an energy-efficient as well as an
environmentally friendly lamp that is able to be utilized for primarily home light. The RGB light
module is made up of an ESP8266 microcontroller, which controls the brightness of the RGB LED
lamp using pulse width modulation ( The RGB lamp itself is controlled by the user using
an Android-based application, which communicates via Wi-Fi (Wireless Fidelity) with a central host
that passes the command to the lamp device through TCP/IP protocol. The RGB lamp that uses
42.8 mA current consumption in an idle condition and 603.1 mA in the active condition for 12 VDC
sup ply voltage.