Wireless AC Power control Circuit
This Project is based on the concept of home automation, but what differentiates it from others is that we can deploy this system on existing lights of houses irrespective of its type i.e. CFL, LED, Fluorescent, etc. And not only on lights but Fans as well. This solution can be deployed to the existing circuitry of the house and has a very optimal cost of installation. Working with the ESP Wi-Fi Modules to build a smart Wi-Fi socket that enables users to control their AC loads wirelessly through smartphones. While products like these are already available in the market, like the popular Moko WIFI Smart Plug or Sonoff, they are a bit costlier. So, this project enables us to build a smart plug using ESP8266 WIFI module which is affordable and cost convenient. The device that we built can easily be plugged into any exiting AC socket and then on the other point you can connect the actual load simply by plugging it into this socket on the device. later just keep the main switch off the socket always on, and hence we can control load directly from the Smartphone.