SMART THERMAL HYGROMTER
Temperature can be defined as the measure of warmth or coldness of an object or substance with reference to some standard value. In a similar way, humidity can be defined as the measure of the amount of water vapor present in the air.
To measure both of these a device called Smart Thermal Hygrometer is developed to perform thermal screening and hence both temperature and humidity can be measured without having physical contact.
The necessity of designing a thermal hygrometer with thermal screening is that in few cases like measuring the temperature of certain machines like DC motor it is not possible to have physical contact, and also nowadays Corona has led to the circumstances that the temperature has to be measured with no physical contact.