RGB LED-IoT Based
In this work, an RGB light-emitting diode (LED) lamp was designed for a smart home system. The ultimate aim of this research is to produce an energy-efficient as well as an environmentally-friendly lamp that is able to be utilized for primarily home activities. The RGB lamp device is made up of an ESP8266 microcontroller, which controls RGB LED lamp using pulse-width modulation (PWM). The RGB lamp device itself is controlled by the user using an Android-based application, which communicates using Wi-Fi (Wireless Fidelity) with a central host that passes the command to the lamp device through TCP/IP protocol. The end result of this work is an RGB lamp that uses 42.8 mA current consumption in an idle condition and 603.1 mA in the active condition for 5 VDC supply voltage.