Electronic Dice-Logic circuit based
This project includes a 555-Timer, IC 4017, IC 4511 based analogue & digital dice game using an electronic digital dice with the help of LEDs & 7 segment display. The game design is a simple electronic circuit with a push button that can be used as a normal dice in games. The dice works by switching Light Emitting Diodes (LEDs) on and off just like the pattern of dots found on a traditional dice & also the number is displayed in the 7-segment display, which makes this game both analogue & digital. We show how 555-timer may be more or less dependable and suggest some improvements to standard designs. The paper includes many recommendations for developers and purchasers.